+++ app/src/main/java/com/example/coursesapp/features/authorization/di/AuthModule.kt
@@ -1,0 +1,2 @@
+package com.example.coursesapp.features.authorization.di
+
