+++ app/src/main/java/com/example/coursesapp/features/home/ui/HomeScreen.kt
@@ -1,4 +1,4 @@
-package com.example.coursesapp.home
+package com.example.coursesapp.features.home.ui
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
@@ -32,7 +32,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.example.coursesapp.R
 import com.example.coursesapp.core.composable.CourseTextField
-import com.example.coursesapp.home.items.CourseItem
+import com.example.coursesapp.features.home.domain.items.CourseItem
 import com.example.coursesapp.ui.theme.BasicGreen
 import com.example.coursesapp.ui.theme.CoursesAppTheme
 import com.example.coursesapp.ui.theme.DarkGrey
