+++ app/src/main/java/com/example/coursesapp/main/navigation/Navigation.kt
@@ -24,13 +24,13 @@ import androidx.navigation.compose.composable
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
 import com.example.coursesapp.R
-import com.example.coursesapp.authorization.AuthorizationScreen
-import com.example.coursesapp.authorization.navigation.AUTH_ROUTE
-import com.example.coursesapp.home.HomeScreen
-import com.example.coursesapp.home.navigation.HOME_ROUTE
-import com.example.coursesapp.home.navigation.navigateToHomeScreen
-import com.example.coursesapp.onboarding.OnboardingScreen
-import com.example.coursesapp.onboarding.navigation.ONBOARD_ROUTE
+import com.example.coursesapp.features.authorization.ui.AuthorizationScreen
+import com.example.coursesapp.features.authorization.navigation.AUTH_ROUTE
+import com.example.coursesapp.features.home.ui.HomeScreen
+import com.example.coursesapp.features.home.navigation.HOME_ROUTE
+import com.example.coursesapp.features.home.navigation.navigateToHomeScreen
+import com.example.coursesapp.features.onboarding.OnboardingScreen
+import com.example.coursesapp.features.onboarding.navigation.ONBOARD_ROUTE
 import com.example.coursesapp.ui.theme.BasicGreen
 import com.example.coursesapp.ui.theme.BasicGrey
 import com.example.coursesapp.ui.theme.DarkGrey
